!include/posix-socket.h: change socklen_t to u32
OpenGroup:
To forestall portability problems, it is recommended that applications
not use values larger than 2^31 - 1 for the socklen_t type.

TASK: RTOS-1367

+++ include/posix-socket.h
@@ -74,7 +74,7 @@
 #define SHUT_RDWR 2
 
 
-typedef size_t socklen_t;
+typedef __u32 socklen_t;
 typedef __u16 sa_family_t;
 
 struct sockaddr {
